demo-apps/backends/python/requirements.txt
# FastAPI框架和依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库相关
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.0

# 验证和安全
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.0

# HTTP客户端和工具
httpx==0.25.2
python-dateutil==2.8.2

# 开发和测试工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 代码质量工具（与pyproject.toml配置对应）
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3